Refactor project into modular package
